Dialogue for navigation
Made a system for rooms and items so you can click an  itemm then get back to the room description. The items are now in their own dialogue file and each throws back to the room your in when its done.

I don't know how elegant it is but it works.

There's also an issue where the room loads before the dailogue bubble at launch so I had to create a signal for it. Used the persistence system for that.